requirements.txt
# Core Framework
fastapi==0.111.0
uvicorn[standard]==0.30.1
pydantic==2.7.4
pydantic-settings==2.3.4
email-validator==2.2.0

# Database
sqlalchemy[asyncio]==2.0.31
asyncpg==0.29.0
alembic==1.13.1
greenlet==3.0.3

# Authentication
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
bcrypt==4.1.3

# File Handling
python-multipart==0.0.9
aiofiles==24.1.0
PyPDF2==3.0.1
boto3==1.34.131
botocore==1.34.131

# GenAI/LLM
httpx==0.27.0
openai==1.35.7

# ML/Recommendation
scikit-learn==1.5.1
numpy==1.26.4

# Background Tasks
celery[redis]==5.4.0
redis==5.0.7
